--- pkg/cli/armtemplate/providers/doc.go
@@ -1,11 +1,0 @@
-// ------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-// ------------------------------------------------------------
-
-package providers
-
-/*
-Package providers contains the code to allow the DeploymentEvaluator
-to interact with resources provided by an extension like Kubernetes.
-*/
