--- pkg/cli/armtemplate/testdata/arm/Microsoft.CustomProviders-resourceProviders-Application-HttpRouteradiusv3-azure-resources-container-httpbinding-backend.json
@@ -1,14 +1,0 @@
-{
-    "id": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend",
-    "type": "Microsoft.CustomProviders/resourceProviders/Application/HttpRoute",
-    "apiVersion": "2018-09-01-preview",
-    "name": "radiusv3/azure-resources-container-httpbinding/backend",
-    "dependsOn": [
-        "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding"
-    ],
-    "body": {
-        "properties": {
-            "port": 80
-        }
-    }
-}
