--- pkg/cli/armtemplate/testdata/arm/Microsoft.CustomProviders-resourceProviders-Application-HttpRouteradiusv3-azure-resources-container-httpbinding-frontend.json
@@ -1,14 +1,0 @@
-{
-    "id": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/frontend",
-    "type": "Microsoft.CustomProviders/resourceProviders/Application/HttpRoute",
-    "apiVersion": "2018-09-01-preview",
-    "name": "radiusv3/azure-resources-container-httpbinding/frontend",
-    "dependsOn": [
-        "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding"
-    ],
-    "body": {
-        "properties": {
-            "port": 80
-        }
-    }
-}
