--- pkg/cli/armtemplate/testdata/armevaluated/Microsoft.CustomProviders-resourceProviders-Application-Containerradiusv3-azure-resources-container-httpbinding-frontend.json
@@ -1,34 +1,0 @@
-{
-    "id": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/Container/frontend",
-    "type": "Microsoft.CustomProviders/resourceProviders/Application/Container",
-    "apiVersion": "2018-09-01-preview",
-    "name": "radiusv3/azure-resources-container-httpbinding/frontend",
-    "dependsOn": [
-        "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding",
-        "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend",
-        "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/frontend"
-    ],
-    "body": {
-        "properties": {
-            "connections": {
-                "backend": {
-                    "kind": "Http",
-                    "source": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend"
-                }
-            },
-            "container": {
-                "env": {
-                    "SERVICE__BACKEND__PORT": 80,
-                    "ANOTHER__PORT": 80
-                },
-                "image": "rynowak/frontend:0.5.0-dev",
-                "ports": {
-                    "web": {
-                        "containerPort": 80,
-                        "provides": "/subscriptions/test-sub/resourceGroups/test-group/providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/frontend"
-                    }
-                }
-            }
-        }
-    }
-}
