--- pkg/cli/armtemplate/testdata/frontend-backend.json
@@ -1,110 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "metadata": {
-    "_generator": {
-      "name": "bicep",
-      "version": "0.4.721.5481",
-      "templateHash": "12305972426212226061"
-    }
-  },
-  "functions": [],
-  "parameters": {
-    "frontendRoute": {
-      "type": "string"
-    },
-    "backendRoute": {
-      "type": "string",
-      "defaultValue": "backend"
-    }
-  },
-  "resources": [
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application/HttpRoute",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}/{2}', 'radiusv3', 'azure-resources-container-httpbinding', parameters('frontendRoute'))]",
-      "properties": {
-        "port": 80
-      },
-      "dependsOn": [
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application', 'radiusv3', 'azure-resources-container-httpbinding')]"
-      ]
-    },
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application/Container",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}/{2}', 'radiusv3', 'azure-resources-container-httpbinding', 'frontend')]",
-      "properties": {
-        "connections": {
-          "backend": {
-            "kind": "Http",
-            "source": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]"
-          }
-        },
-        "container": {
-          "image": "rynowak/frontend:0.5.0-dev",
-          "ports": {
-            "web": {
-              "containerPort": 80,
-              "provides": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('frontendRoute'))]"
-            }
-          },
-          "env": {
-            "SERVICE__BACKEND__PORT": "[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))).port]",
-            "ANOTHER__PORT": "[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute')))['port']]"
-          }
-        }
-      },
-      "dependsOn": [
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application', 'radiusv3', 'azure-resources-container-httpbinding')]",
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]",
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('frontendRoute'))]"
-      ]
-    },
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application/HttpRoute",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}/{2}', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]",
-      "properties": {
-        "port": 80
-      },
-      "dependsOn": [
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application', 'radiusv3', 'azure-resources-container-httpbinding')]"
-      ]
-    },
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application/Container",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}/{2}', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]",
-      "properties": {
-        "container": {
-          "image": "rynowak/backend:0.5.0-dev",
-          "ports": {
-            "web": {
-              "containerPort": 81,
-              "provides": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]"
-            }
-          }
-        }
-      },
-      "dependsOn": [
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application', 'radiusv3', 'azure-resources-container-httpbinding')]",
-        "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]"
-      ]
-    },
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}', 'radiusv3', 'azure-resources-container-httpbinding')]"
-    }
-  ],
-  "outputs": {
-    "test": {
-      "type": "string",
-      "value": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', 'frontendRoute')]",
-      "metadata": {
-        "resourceType": "radius.dev/Application/HttpRoute@v1alpha3"
-      }
-    }
-  }
-}
