--- pkg/cli/armtemplate/testdata/frontend-backend/Application-azure-resources-container-httpbinding.json
@@ -1,27 +1,0 @@
-{
-    "apiVersion": "radius.dev/v1alpha3",
-    "kind": "Application",
-    "metadata": {
-        "annotations": {
-            "radius.dev/application": "azure-resources-container-httpbinding"
-        },
-        "labels": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "app.kubernetes.io/name": "azure-resources-container-httpbinding",
-            "app.kubernetes.io/managed-by": "radius-rp"
-        },
-        "name": "azure-resources-container-httpbinding",
-        "namespace": "default"
-    },
-    "spec": {
-        "application": "azure-resources-container-httpbinding",
-        "template": {
-            "id": "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding",
-            "type": "Microsoft.CustomProviders/resourceProviders/Application",
-            "apiVersion": "2018-09-01-preview",
-            "name": "radiusv3/azure-resources-container-httpbinding",
-            "dependsOn": [],
-            "body": {}
-        }
-    }
-}
