--- pkg/cli/armtemplate/testdata/frontend-backend/Container-backend.json
@@ -1,48 +1,0 @@
-{
-    "apiVersion": "radius.dev/v1alpha3",
-    "kind": "Container",
-    "metadata": {
-        "annotations": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "backend",
-            "radius.dev/resource-type": "Container"
-        },
-        "labels": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "backend",
-            "radius.dev/resource-type": "Container",
-            "app.kubernetes.io/name": "backend",
-            "app.kubernetes.io/part-of": "azure-resources-container-httpbinding",
-            "app.kubernetes.io/managed-by": "radius-rp"
-        },
-        "name": "azure-resources-container-httpbinding-backend",
-        "namespace": "default"
-    },
-    "spec": {
-        "application": "azure-resources-container-httpbinding",
-        "resource": "backend",
-        "template": {
-            "id": "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/Container/backend",
-            "type": "Microsoft.CustomProviders/resourceProviders/Application/Container",
-            "apiVersion": "2018-09-01-preview",
-            "name": "radiusv3/azure-resources-container-httpbinding/backend",
-            "dependsOn": [
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding",
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend"
-            ],
-            "body": {
-                "properties": {
-                    "container": {
-                        "image": "rynowak/backend:0.5.0-dev",
-                        "ports": {
-                            "web": {
-                                "containerPort": 81,
-                                "provides": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]"
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
