--- pkg/cli/armtemplate/testdata/frontend-backend/Container-frontend.json
@@ -1,59 +1,0 @@
-{
-    "apiVersion": "radius.dev/v1alpha3",
-    "kind": "Container",
-    "metadata": {
-        "annotations": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "frontend",
-            "radius.dev/resource-type": "Container"
-        },
-        "labels": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "frontend",
-            "radius.dev/resource-type": "Container",
-            "app.kubernetes.io/name": "frontend",
-            "app.kubernetes.io/part-of": "azure-resources-container-httpbinding",
-            "app.kubernetes.io/managed-by": "radius-rp"
-        },
-        "name": "azure-resources-container-httpbinding-frontend",
-        "namespace": "default"
-    },
-    "spec": {
-        "application": "azure-resources-container-httpbinding",
-        "resource": "frontend",
-        "template": {
-            "id": "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/Container/frontend",
-            "type": "Microsoft.CustomProviders/resourceProviders/Application/Container",
-            "apiVersion": "2018-09-01-preview",
-            "name": "radiusv3/azure-resources-container-httpbinding/frontend",
-            "dependsOn": [
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding",
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend",
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/frontend"
-            ],
-            "body": {
-                "properties": {
-                    "connections": {
-                        "backend": {
-                            "kind": "Http",
-                            "source": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))]"
-                        }
-                    },
-                    "container": {
-                        "env": {
-                            "SERVICE__BACKEND__PORT": "[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute'))).port]",
-                            "ANOTHER__PORT": "[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('backendRoute')))['port']]"
-                        },
-                        "image": "rynowak/frontend:0.5.0-dev",
-                        "ports": {
-                            "web": {
-                                "containerPort": 80,
-                                "provides": "[resourceId('Microsoft.CustomProviders/resourceProviders/Application/HttpRoute', 'radiusv3', 'azure-resources-container-httpbinding', parameters('frontendRoute'))]"
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
