--- pkg/cli/armtemplate/testdata/frontend-backend/HttpRoute-backend.json
@@ -1,39 +1,0 @@
-{
-    "apiVersion": "radius.dev/v1alpha3",
-    "kind": "HttpRoute",
-    "metadata": {
-        "annotations": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "backend",
-            "radius.dev/resource-type": "HttpRoute"
-        },
-        "labels": {
-            "radius.dev/application": "azure-resources-container-httpbinding",
-            "radius.dev/resource": "backend",
-            "radius.dev/resource-type": "HttpRoute",
-            "app.kubernetes.io/name": "backend",
-            "app.kubernetes.io/part-of": "azure-resources-container-httpbinding",
-            "app.kubernetes.io/managed-by": "radius-rp"
-        },
-        "name": "azure-resources-container-httpbinding-backend",
-        "namespace": "default"
-    },
-    "spec": {
-        "application": "azure-resources-container-httpbinding",
-        "resource": "backend",
-        "template": {
-            "id": "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding/HttpRoute/backend",
-            "type": "Microsoft.CustomProviders/resourceProviders/Application/HttpRoute",
-            "apiVersion": "2018-09-01-preview",
-            "name": "radiusv3/azure-resources-container-httpbinding/backend",
-            "dependsOn": [
-                "/subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/azure-resources-container-httpbinding"
-            ],
-            "body": {
-                "properties": {
-                    "port": 80
-                }
-            }
-        }
-    }
-}
