--- pkg/cli/armtemplate/testdata/k8s/conn/output.json
@@ -1,17 +1,0 @@
-{
-    "connectionString": {
-        "properties": {
-            "metadata": {
-                "name": "connectionString",
-                "namespace": "default",
-                "labels": {
-                    "format": "custom"
-                }
-            },
-            "data": {
-                "conn": "redis-master.default.svc.cluster.local,password=awesome",
-                "port": "6379"
-            }
-        }
-    }
-}
