--- pkg/cli/armtemplate/testdata/k8s/conn/secret.yaml
@@ -1,8 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: redis
-  namespace: default
-type: Opaque
-data:
-  redisPassword: YXdlc29tZQ==
