--- pkg/cli/armtemplate/testdata/k8s/conn/service.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: redis-master
-  namespace: default
-spec:
-  type: ClusterIP
-  ports:
-  - name: tcp-redis
-    port: 6379
-    protocol: TCP
-    targetPort: redis
