--- pkg/cli/armtemplate/testdata/k8s/missing-resource/errors.json
@@ -1,3 +1,0 @@
-[
-   "services \"redis-master\" not found"
-]
