--- pkg/cli/armtemplate/testdata/managed-secret/redis.yaml
@@ -1,32 +1,0 @@
-apiVersion: radius.dev/v1alpha3
-kind: RedisComponent
-metadata:
-  annotations:
-    radius.dev/application: kubernetes-resources-redis
-    radius.dev/resource: redis
-    radius.dev/resource-type: redislabs.com.RedisCache
-  labels:
-    app.kubernetes.io/managed-by: radius-rp
-    app.kubernetes.io/name: redis
-    app.kubernetes.io/part-of: kubernetes-resources-redis
-    radius.dev/application: kubernetes-resources-redis
-    radius.dev/resource: redis
-    radius.dev/resource-type: redislabs.com.RedisCache
-  name: kubernetes-resources-redis
-  namespace: default
-spec:
-  application: kubernetes-resources-redis
-  resource: redis
-  template:
-    apiVersion: 2018-09-01-preview
-    body:
-      properties:
-        host: my-redis.com
-        port: 1234
-        secrets:
-          connectionString: '*****'
-          password: '*****'
-    dependsOn: []
-    id: /subscriptions//resourceGroups//providers/Microsoft.CustomProviders/resourceProviders/radiusv3/Application/kubernetes-resources-redis/redislabs.com.RedisCache/redis
-    name: radiusv3/kubernetes-resources-redis/redis
-    type: Microsoft.CustomProviders/resourceProviders/Application/redislabs.com.RedisCache
