--- pkg/cli/armtemplate/testdata/managed-secret/secret.yaml
@@ -1,2 +1,0 @@
-connectionString: connectionString
-password: password
