--- pkg/cli/armtemplate/testdata/managed-secret/template.json
@@ -1,27 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "metadata": {
-    "_generator": {
-      "name": "bicep",
-      "version": "0.4.1004.46306",
-      "templateHash": "5971979958412524905"
-    }
-  },
-  "functions": [],
-  "resources": [
-    {
-      "type": "Microsoft.CustomProviders/resourceProviders/Application/redislabs.com.RedisCache",
-      "apiVersion": "2018-09-01-preview",
-      "name": "[format('{0}/{1}/{2}', 'radiusv3', 'kubernetes-resources-redis-unmanaged', 'redis')]",
-      "properties": {
-        "host": "my-redis.com",
-        "port": 1234,
-        "secrets": {
-          "connectionString": "connectionString",
-          "password": "password"
-        }
-      }
-    }
-  ]
-}
